+++ inferno-core/app/Main.hs
@@ -1,13 +1,18 @@
 {-# LANGUAGE LambdaCase #-}
-{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
 
 module Main where
 
-import Data.Bifunctor (bimap)
+import Data.Bifunctor (bimap, second)
 import qualified Data.Map as Map
 import qualified Data.Text.IO as Text
-import Inferno.Core (Interpreter (..), mkInferno)
-import Inferno.Module.Prelude (builtinModules)
+import Inferno.Eval (runEvalM)
+import Inferno.Infer (inferExpr)
+import Inferno.Infer.Pinned (pinExpr)
+import Inferno.Module.Prelude (baseOpsTable, builtinModules, builtinModulesOpsTable, builtinModulesPinMap, builtinModulesTerms)
+import Inferno.Parse (parseExpr)
+import Inferno.Parse.Commented (insertCommentsIntoExpr)
 import Inferno.Types.VersionControl (pinnedToMaybe)
 import Inferno.Utils.Prettyprinter (showPretty)
 import Options.Applicative
@@ -30,13 +35,14 @@ import Options.Applicative
 import System.Exit (exitFailure)
 import System.IO (hPrint, stderr)
 
-data CliArgs = CliArgs {file :: String, typecheck :: Bool}
+data CliArgs = CliArgs {file :: String, typecheck :: Bool, parse :: Bool}
 
 cliargs :: Parser CliArgs
 cliargs =
   CliArgs
     <$> argument str (metavar "FILE" <> help "Input file path")
     <*> switch (long "typecheck" <> short 't' <> help "Only run type inference")
+    <*> switch (long "parse" <> short 'p' <> help "Only run parser")
 
 main :: IO ()
 main = do
@@ -50,21 +56,43 @@ main = do
   args <- execParser opts
 
   src <- Text.readFile $ file args
-  Interpreter {evalExpr, defaultEnv, parseAndInfer} <-
-    mkInferno builtinModules [] :: IO (Interpreter IO ())
-  case parseAndInfer src of
+  let customTypes = []
+  let prelude = builtinModules @IO @()
+  let defaultEnv = builtinModulesTerms prelude
+  -- parse
+  case parseExpr (baseOpsTable prelude) (builtinModulesOpsTable prelude) customTypes src of
     Left err -> do
       hPrint stderr err
       exitFailure
-    Right (ast, ty, _, _) -> do
-      if typecheck args
+    Right (ast, _comments) ->
+      if parse args
         then do
-          putStrLn "Inferred type:"
-          showPretty ty
+          let ast' = insertCommentsIntoExpr _comments ast
+          putStrLn "Parsed Expr:"
+          print $ second (const ()) ast'
+          putStrLn "Pretty:"
+          showPretty ast'
         else do
-          let ast' = bimap pinnedToMaybe (const ()) ast
-          evalExpr defaultEnv Map.empty ast' >>= \case
+          -- pin free variables to builtin prelude function hashes
+          case pinExpr (builtinModulesPinMap prelude) ast of
             Left err -> do
               hPrint stderr err
               exitFailure
-            Right res -> showPretty res
+            Right pinnedAST ->
+              -- typecheck
+              case inferExpr prelude pinnedAST of
+                Left err -> do
+                  hPrint stderr err
+                  exitFailure
+                Right (ast', ty, _tyMap) ->
+                  if typecheck args
+                    then do
+                      putStrLn "Inferred type:"
+                      showPretty ty
+                    else do
+                      let ast'' = bimap pinnedToMaybe (const ()) ast'
+                      runEvalM defaultEnv Map.empty ast'' >>= \case
+                        Left err -> do
+                          hPrint stderr err
+                          exitFailure
+                        Right res -> showPretty res