+++ inferno-core/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Revision History for inferno-core
 *Note*: we use https://pvp.haskell.org/ (MAJOR.MAJOR.MINOR.PATCH)
 
+## 0.11.4.0 -- 2024-04-02
+* Add Inferno documentation and language guide to Inferno.Docs
+
+## 0.11.3.1 -- 2024-04-01
+* Fix record field order in NonExhaustivePatternMatch error
+
+## 0.11.3.0 -- 2024-04-01
+* Add --parse CLI arg and allow trailing commas in parser
+
 ## 0.11.2.0 -- 2024-03-26
 * Add record pattern matching